rename applets: voice-agent→conversational-agent, conversational-agent→wake-command-agent
Mirrors the same rename applied to corti-sdk-sandbox:
- "Voice agent" is now "Conversational agent" (always-on turn-based agent with speculative prefetch)
- "Conversational agent" is now "Wake-command agent" (wake-phrase gated chat with threaded memory)
- Updates display names, route IDs, directory names, file names, component symbols,
  exported constants, and localStorage key namespaces in both applets
- Registry, imports, and README headings updated to match
