--- sdk/typescript/applets/src/applets/voice-agent/config.ts
@@ -1,10 +1,0 @@
-import type { Corti } from "@corti/sdk";
-
-export function buildVoiceConfig(primaryLanguage: string): Corti.TranscribeConfig {
-  return {
-    primaryLanguage,
-    interimResults: true,
-    automaticPunctuation: false,
-    audioEvents: { enabled: true },
-  };
-}
