+++ sdk/typescript/applets/src/applets/wake-command-agent/config.test.ts
