+++ sdk/typescript/applets/src/applets/wake-command-agent/model.test.ts
